--- Temp.java
@@ -1,74 +1,0 @@
-import java.util.ArrayList;
-import java.util.List;
-
-public class Temp {
-
-   
-    public static void main(String[] args) {
-
-       List strlist = new ArrayList(); //list is interface so its proper object cannot be formed
-       
-      //If we want the list to contain specific type of elements then put it in <> on left side 
-      // List <String> strlist = new ArrayList<>();  
-
-      
-
-        strlist.add("Indian"); // this is the 0th index;
-        strlist.add("league");// this is 1st index
-       strlist.add(1);// it can add different datatypes as well if we dont mention <datatype> during object creation !!!
-
-        System.out.println(strlist.get(0)); //accessing the element
-
-        for(int i =0;i<strlist.size();i++){     // accessing elements via forloop
-            System.out.print(strlist.get(i) + "  ");
-        }
-
-        System.out.println("\n \n");
-
-
-        strlist.add(1,"premier"); //adding the element in between the list
-        
-        for(int i=0;i<strlist.size();i++){
-            System.out.print( i + ":" + strlist.get(i) + "  ");
-        }
-        
-        
-        
-        System.out.println("\n\n" + " after removing : "+"Indian");
-        strlist.remove("Indian");
-        for(int i=0;i<strlist.size();i++){
-            System.out.print(i + ":"  + strlist.get(i)+ "   ");
-        }      
-
-        System.out.println("\n\n" + " removal of element by index: ");
-        strlist.remove(2); //deleting league which is at index 1
-        for(int i=0;i<strlist.size();i++){
-            System.out.print(i +  ":" + strlist.get(i)+  "  ");
-        }
-
-
-        // for replacing the element in list
-        System.out.println("\n\n replacing the element in list :-");
-        strlist.set(1,"Nepal");      //replaced
-
-        for(int i=0;i<strlist.size();i++){
-            System.out.print(i + ":" + strlist.get(i) + "  ");
-        }
-
-      
-        System.out.println(" \n\n " + "size of the list is : " + strlist.size());
-
-
-        System.out.println("list.contains(object): checks weather the list contains the object or not");
-
-        if(strlist.contains("Nepal")){
-            System.out.println("Nepal exists");
-            System.out.println("at index = " + strlist.indexOf("Nepal"));
-        }
-
-
- 
-    }
-
-    
-}
